requirements.txt
pyspark==3.5.0
delta-spark==3.2.0
pandas==2.2.3
pyarrow==19.0.1

pytest==8.3.0
black==25.1.0
mypy==1.15.0
pre-commit==4.3.0
isort==6.1.0
flake8==7.3.0
